// src/utils/enum.ts
export enum EventStatusType {
  COMPLETED = "COMPLETED",
  CANCELED = "CANCELED",
  PLANNING = "PLANNING",
}

export enum EventStatusColor {
  COMPLETED = "bg-green-200",
  CANCELED = "bg-red-200",
  PLANNING = "bg-blue-200",
}

export enum EventStatusColorCode {
  COMPLETED = "#A7F3D0",
  CANCELED = "#FECACA",
  PLANNING = "#BFDBFE",
}

export enum GuestStatusColor {
  ATTEND = "bg-fiesta-200",
  DECLINE = "bg-slate-900",
  WAITING = "bg-gray-300",
}

export enum GuestStatusColorCode {
  ATTEND = "#ddd6fe",
  DECLINE = "#020617",
  WAITING = "#d1d5db",
}